Process review comments. Avoid duplication.
