+++ packages/ui/src/forum/hooks/useForumCategories.ts
@@ -1,12 +1,11 @@
-import { useGetForumCategoriesQuery, useGetArchivedForumCategoriesQuery } from '@/forum/queries'
-import { asForumCategory, asArchivedForumCategory, CategoryStatusType } from '@/forum/types'
+import { useGetForumCategoriesQuery } from '@/forum/queries'
+import { asForumCategory, CategoryStatusType } from '@/forum/types'
 
 interface Props {
   isRoot?: boolean
 }
 
 export const ActiveStatus: CategoryStatusType = 'CategoryStatusActive'
-export const ArchivedStatus: CategoryStatusType = 'CategoryStatusArchived'
 
 export const useForumCategories = ({ isRoot }: Props = {}) => {
   const { loading, data } = useGetForumCategoriesQuery({
@@ -20,20 +19,5 @@ export const useForumCategories = ({ isRoot }: Props = {}) => {
   return {
     isLoading: loading,
     forumCategories: data?.forumCategories.map(asForumCategory),
-  }
-}
-
-export const useArchivedForumCategories = () => {
-  const { loading, data } = useGetArchivedForumCategoriesQuery({
-    variables: {
-      where: {
-        status_json: { isTypeOf_eq: ArchivedStatus },
-      },
-    },
-  })
-
-  return {
-    isLoading: loading,
-    forumCategories: data?.forumCategories.map(asArchivedForumCategory),
   }
 }
